+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Utxow.hs
@@ -39,8 +39,8 @@ import Cardano.Ledger.Alonzo.Rules.Utxos (AlonzoUtxosPredFailure)
 import Cardano.Ledger.Alonzo.Scripts (plutusScriptLanguage, toAsItem, toAsIx)
 import Cardano.Ledger.Alonzo.Tx (hashScriptIntegrity)
 import Cardano.Ledger.Alonzo.TxWits (
-  unRedeemers,
-  unTxDats,
+  unRedeemersL,
+  unTxDatsL,
  )
 import Cardano.Ledger.Alonzo.UTxO (
   AlonzoEraUTxO (..),
@@ -237,7 +237,7 @@ missingRequiredDatums utxo tx = do
   let txBody = tx ^. bodyTxL
       scriptsProvided = getScriptsProvided utxo tx
       (inputHashes, txInsNoDataHash) = getInputDataHashesTxBody utxo txBody scriptsProvided
-      txHashes = domain (unTxDats $ tx ^. witsTxL . datsTxWitsL)
+      txHashes = domain (tx ^. witsTxL . datsTxWitsL . unTxDatsL)
       unmatchedDatumHashes = eval (inputHashes ➖ txHashes)
       allowedSupplementalDataHashes = getSupplementalDataHashes utxo txBody
       supplimentalDatumHashes = eval (txHashes ➖ inputHashes)
@@ -274,7 +274,7 @@ hasExactSetOfRedeemers tx (ScriptsProvided scriptsProvided) (AlonzoScriptsNeeded
         ]
       (extraRdmrs, missingRdmrs) =
         extSymmetricDifference
-          (Map.keys $ unRedeemers $ tx ^. witsTxL . rdmrsTxWitsL)
+          (Map.keys $ tx ^. witsTxL . rdmrsTxWitsL . unRedeemersL)
           id
           redeemersNeeded
           fst