+++ eras/alonzo/impl/CHANGELOG.md
@@ -24,6 +24,8 @@
 * Stop re-exporting deprecated `AuxiliaryDataHash` from `Cardano.Ledger.Alonzo.TxAuxData`
 * Deprecate `Alonzo` type synonym
 * Remove crypto parametrization from `AlonzoEra`
+* Add `unRedeemersL`, `unTxDatsL`
+* Deprecate `lookupRedeemer`, `nullRedeemers`, `nullDats`
 
 ### `testlib`